requirements.txt
snowflake-connector-python>=3.0.0
snowflake-snowpark-python>=1.0.0
snowflake-ml-python>=1.0.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0
cryptography>=3.4.0
pyarrow>=10.0.0
pytest>=7.0.0 